hbi,blockchain: fix create tx id
Transaction ID is created base on protobuf.Tx payload. Since when we
added ExternalAddress to protobuf.Tx payload, it's a map, so the order
of key-element is not specified. So when we marshal json it, the data
return can be different between runs.

To fix it, we can exclude ExternalAddress from tx payload.